+++ Домашнее задание №1. Диаграммы классов/src/INotifyer.cs
@@ -1,5 +1,5 @@
 namespace Homework_1.INotifyer;
 
 public interface INotifyer {
-    public void Notify(int balance);
+    public void Notify(decimal balance);
 }
